Issue #164: Refactoring for more object orientedness.
Add profile_helper class.
Move aggregate functions from profile to profile_helper.
Add PHPDoc.
