+++ classes/processor.php
@@ -19,6 +19,9 @@ namespace tool_excimer;
 /**
  * Interface for classes that define processors for the samples.
  *
+ * A processor takes the samples generated by the profiler manager, and
+ * creates the profile(s) to be saved to disk.
+ *
  * @package   tool_excimer
  * @author    Jason den Dulk <jasondendulk@catalyst-au.net>
  * @copyright 2022, Catalyst IT
@@ -30,7 +33,7 @@ interface processor {
      *
      * @param manager $manager The profiler manager object
      */
-    public function init(manager $manager): void;
+    public function init(manager $manager);
 
     /**
      * Gets the minimum duration required for a profile to be saved, as seconds.
